feat(swarmplot): move stories to the dedicated storybook workspace
